update: initializer, move initializer types to models add: environment & oauth models
