+++ npm/react/libs/core/src/models/environment.ts
@@ -1,0 +1,11 @@
+import type { OAuth } from './oauth';
+
+export interface Environment {
+  production: boolean;
+  endpoints: Record<string, Endpoint>;
+  oauth: OAuth;
+}
+
+export interface Endpoint {
+  url: string;
+}
