+++ npm/react/libs/core/src/models/oauth.ts
@@ -1,0 +1,7 @@
+export interface OAuth {
+  authority: string;
+  clientId: string;
+  scope: string;
+  responseType: 'password' | 'code';
+  redirectUri: string;
+}
